moved ssl/curl code to separate file; now it configured both libcurl and openssl; added defaults to libcurl static install
