+++ src/daemon/config/netdata-conf-ssl.h
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef NETDATA_NETDATA_CONF_SSL_H
+#define NETDATA_NETDATA_CONF_SSL_H
+
+void netdata_conf_ssl(void);
+
+#endif //NETDATA_NETDATA_CONF_SSL_H
