+++ src/daemon/config/netdata-conf-web.c
@@ -159,7 +159,5 @@ void netdata_conf_web_security_init(void) {
 
     tls_version    = inicfg_get(&netdata_config, CONFIG_SECTION_WEB, "tls version",  "1.3");
     tls_ciphers    = inicfg_get(&netdata_config, CONFIG_SECTION_WEB, "tls ciphers",  "none");
-
-    netdata_ssl_initialize_openssl();
 }
